added new socket events
